implementation of basic functions for a python wva library
This is an initial first cut of a library for talking to the WVA,
which will eventually be a target for release as open-source.
Right now, the functionality included is limited and not well documented
as it is likely to evolve quickly for a little while until it
matures.

The README includes basic ideas for the functionality that will
eventually be exposed by the library.  The library will target
Python 2.7+