http: add exception hierarchy and extract http code from the core
Smaller modules tend to be easier to maintain and test.  They also
make dependencies more obvious.  This change separates out the http
code from the wva core more clearly.  As part of this, rather than
aliasing the methods provided by the http client, we now just
provide a method to get a client reference.  This seems like a reasonable
hurdle to put in place for low-level http access.