events/subscriptions: add support for subscriptions and event streaming
This commit adds initial support for CRUD operaitons on subscriptions
and for receiving callbacks for events containing vehicle data
from the WVA.

In addition, new CLI commands were added that exposes this functionality.
More polish may be added in the future in order to allow for subscriptions
to be created directly from a WVAVehicleDataElement.