graph: add ability to graph incoming stream data
This commit adds a "cool demo" for those individuals who have
matplotlib installed (it is not added on as a dependency as it
is a large dependency).  Usage (having previously created
subscriptions) looks like this:

    $ wva subscriptions graph EngineSpeed,VehicleSpeed --seconds 120

Currently, all data is plotted on the same set of axes which can be
somewhat confusing, but you can make it work.  If users want something
too different, they can just use this as an example and implement
their own functionality.